refactor: 이미 발급받은 쿠폰일떄 사용자에게 바로 응답 주기 (사용자 편의성 제공)
